requirements.txt
chainlit
cohere
openai
tiktoken
python-dotenv
langchain
langchain-openai
pymupdf
qdrant-client
duckduckgo-search